Add "Start Here" section with video content
Added a new section to the landing page using a "markdown" section type, in order to add a video. In the future, we'll want to remove the iframe and convert this to Markdown instead of HTML. Unfortunately, we couldn't do that today because the landing page pipeline doesn't support any of our docs-specific Markdown extensions ;( Once that is resolved, we can fix this.