M11126: Possible mistake in title
Hello, @cshoe,

Translator has reported possible source content issue. According to page https://docs.microsoft.com/es-es/nodejs/api/overview/azure/?view=azure-node-2.0.0, it might look like another Azure service, but we think it is a general term, based in this string: This module includes all of the individual Azure Node.js modules in a single place. 

Please, help to check my proposed file change into the article and help to merge if you agree with fix. If not, please, let me know either if you would like me to fix it in another way within this PR, if you prefer to fix it in another PR, or if I should close this PR as by-design. In case of using another PR, please, let me know of your PR number, so we can confirm and close this PR.

Many thanks in advance.