C26149: HTML content in MD
Hello, @ rloutlaw ,

Localization team has reported source content issue that causes issue for localization. The file is written in HTML. It should be in MD syntax.

Please review the comment tag added and reply with explanation if fix is needed or not. If you make related fix in another PR then share your PR number so we can confirm and close this PR.

Many thanks in advance.